Refactor: move all tests into src/tests
