+++ packages/web-frontend/src/tests/utils/getUniqueViewId.test.js
@@ -5,7 +5,7 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-import { getUniqueViewId, getViewIdFromInfoViewKey } from '../getUniqueViewId';
+import { getUniqueViewId, getViewIdFromInfoViewKey } from '../../utils/getUniqueViewId';
 
 describe('getUniqueViewId', () => {
   describe('getUniqueViewId()', () => {
